+++ modelforge/tests/schnetpack_pain_implementation.py
@@ -1,0 +1,275 @@
+import os
+
+from typing import Callable, Dict, Optional
+
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+import schnetpack.properties as properties
+import schnetpack.nn as snn
+
+
+class PaiNNInteraction(nn.Module):
+    r"""PaiNN interaction block for modeling equivariant interactions of atomistic systems."""
+
+    def __init__(self, n_atom_basis: int, activation: Callable):
+        """
+        Args:
+            n_atom_basis: number of features to describe atomic environments.
+            activation: if None, no activation function is used.
+            epsilon: stability constant added in norm to prevent numerical instabilities
+        """
+        super(PaiNNInteraction, self).__init__()
+        self.n_atom_basis = n_atom_basis
+
+        self.interatomic_context_net = nn.Sequential(
+            snn.Dense(n_atom_basis, n_atom_basis, activation=activation),
+            snn.Dense(n_atom_basis, 3 * n_atom_basis, activation=None),
+        )
+
+    def forward(
+        self,
+        q: torch.Tensor,
+        mu: torch.Tensor,
+        Wij: torch.Tensor,
+        dir_ij: torch.Tensor,
+        idx_i: torch.Tensor,
+        idx_j: torch.Tensor,
+        n_atoms: int,
+    ):
+        """Compute interaction output.
+
+        Args:
+            q: scalar input values
+            mu: vector input values
+            Wij: filter
+            idx_i: index of center atom i
+            idx_j: index of neighbors j
+
+        Returns:
+            atom features after interaction
+        """
+        # inter-atomic
+        x = self.interatomic_context_net(q)
+        xj = x[idx_j]
+        muj = mu[idx_j]
+        x = Wij * xj
+
+        dq, dmuR, dmumu = torch.split(x, self.n_atom_basis, dim=-1)
+        dq = snn.scatter_add(dq, idx_i, dim_size=n_atoms)
+        dmu = dmuR * dir_ij[..., None] + dmumu * muj
+        dmu = snn.scatter_add(dmu, idx_i, dim_size=n_atoms)
+
+        q = q + dq
+        mu = mu + dmu
+
+        return q, mu
+
+
+class PaiNNMixing(nn.Module):
+    r"""PaiNN interaction block for mixing on atom features."""
+
+    def __init__(self, n_atom_basis: int, activation: Callable, epsilon: float = 1e-8):
+        """
+        Args:
+            n_atom_basis: number of features to describe atomic environments.
+            activation: if None, no activation function is used.
+            epsilon: stability constant added in norm to prevent numerical instabilities
+        """
+        super(PaiNNMixing, self).__init__()
+        self.n_atom_basis = n_atom_basis
+
+        self.intraatomic_context_net = nn.Sequential(
+            snn.Dense(2 * n_atom_basis, n_atom_basis, activation=activation),
+            snn.Dense(n_atom_basis, 3 * n_atom_basis, activation=None),
+        )
+        self.mu_channel_mix = snn.Dense(
+            n_atom_basis, 2 * n_atom_basis, activation=None, bias=False
+        )
+        self.epsilon = epsilon
+
+    def forward(self, q: torch.Tensor, mu: torch.Tensor):
+        """Compute intraatomic mixing.
+
+        Args:
+            q: scalar input values
+            mu: vector input values
+
+        Returns:
+            atom features after interaction
+        """
+        ## intra-atomic
+        mu_mix = self.mu_channel_mix(mu)
+        mu_V, mu_W = torch.split(mu_mix, self.n_atom_basis, dim=-1)
+        mu_Vn = torch.sqrt(torch.sum(mu_V**2, dim=-2, keepdim=True) + self.epsilon)
+
+        ctx = torch.cat([q, mu_Vn], dim=-1)
+        x = self.intraatomic_context_net(ctx)
+
+        dq_intra, dmu_intra, dqmu_intra = torch.split(x, self.n_atom_basis, dim=-1)
+        dmu_intra = dmu_intra * mu_W
+
+        dqmu_intra = dqmu_intra * torch.sum(mu_V * mu_W, dim=1, keepdim=True)
+
+        q = q + dq_intra + dqmu_intra
+        mu = mu + dmu_intra
+        return q, mu
+
+
+class PaiNN(nn.Module):
+    """PaiNN - polarizable interaction neural network
+
+    References:
+
+    .. [#painn1] Schütt, Unke, Gastegger:
+    Equivariant message passing for the prediction of tensorial properties and molecular spectra.
+    ICML 2021, http://proceedings.mlr.press/v139/schutt21a.html
+
+    """
+
+    def __init__(
+        self,
+        n_atom_basis: int,
+        n_interactions: int,
+        radial_basis: nn.Module,
+        cutoff_fn: Optional[Callable] = None,
+        activation: Optional[Callable] = F.silu,
+        max_z: int = 100,
+        shared_interactions: bool = False,
+        shared_filters: bool = False,
+        epsilon: float = 1e-8,
+    ):
+        """
+        Args:
+            n_atom_basis: number of features to describe atomic environments.
+                This determines the size of each embedding vector; i.e. embeddings_dim.
+            n_interactions: number of interaction blocks.
+            radial_basis: layer for expanding interatomic distances in a basis set
+            cutoff_fn: cutoff function
+            activation: activation function
+            shared_interactions: if True, share the weights across
+                interaction blocks.
+            shared_interactions: if True, share the weights across
+                filter-generating networks.
+            epsilon: stability constant added in norm to prevent numerical instabilities
+        """
+        super(PaiNN, self).__init__()
+
+        self.n_atom_basis = n_atom_basis
+        self.n_interactions = n_interactions
+        self.cutoff_fn = cutoff_fn
+        self.cutoff = cutoff_fn.cutoff
+        self.radial_basis = radial_basis
+
+        self.embedding = nn.Embedding(max_z, n_atom_basis, padding_idx=0)
+
+        self.share_filters = shared_filters
+
+        if shared_filters:
+            self.filter_net = snn.Dense(
+                self.radial_basis.n_rbf, 3 * n_atom_basis, activation=None
+            )
+        else:
+            self.filter_net = snn.Dense(
+                self.radial_basis.n_rbf,
+                self.n_interactions * n_atom_basis * 3,
+                activation=None,
+            )
+
+        self.interactions = snn.replicate_module(
+            lambda: PaiNNInteraction(
+                n_atom_basis=self.n_atom_basis, activation=activation
+            ),
+            self.n_interactions,
+            shared_interactions,
+        )
+        self.mixing = snn.replicate_module(
+            lambda: PaiNNMixing(
+                n_atom_basis=self.n_atom_basis,
+                activation=activation,
+                epsilon=epsilon,
+            ),
+            self.n_interactions,
+            shared_interactions,
+        )
+
+    def forward(self, inputs: Dict[str, torch.Tensor]):
+        """
+        Compute atomic representations/embeddings.
+
+        Args:
+            inputs (dict of torch.Tensor): SchNetPack dictionary of input tensors.
+
+        Returns:
+            torch.Tensor: atom-wise representation.
+            list of torch.Tensor: intermediate atom-wise representations, if
+            return_intermediate=True was used.
+        """
+        # get tensors from input dictionary
+        atomic_numbers = inputs[properties.Z]
+        r_ij = inputs[properties.Rij]
+        idx_i = inputs[properties.idx_i]
+        idx_j = inputs[properties.idx_j]
+        n_atoms = atomic_numbers.shape[0]
+
+        # compute atom and pair features
+        d_ij = torch.norm(r_ij, dim=1, keepdim=True)
+        dir_ij = r_ij / d_ij
+        phi_ij = self.radial_basis(d_ij)
+        fcut = self.cutoff_fn(d_ij)
+
+        filters = self.filter_net(phi_ij) * fcut[..., None]
+        if self.share_filters:
+            filter_list = [filters] * self.n_interactions
+        else:
+            filter_list = torch.split(filters, 3 * self.n_atom_basis, dim=-1)
+
+        q = self.embedding(atomic_numbers)[:, None]
+        qs = q.shape
+        mu = torch.zeros((qs[0], 3, qs[2]), device=q.device)
+
+        for i, (interaction, mixing) in enumerate(zip(self.interactions, self.mixing)):
+            q, mu = interaction(q, mu, filter_list[i], dir_ij, idx_i, idx_j, n_atoms)
+            q, mu = mixing(q, mu)
+
+        q = q.squeeze(1)
+
+        inputs["scalar_representation"] = q
+        inputs["vector_representation"] = mu
+        print(inputs)
+        # assert False
+        return inputs
+
+
+def setup_painn():
+    import schnetpack as spk
+    from schnetpack.datasets import QM9
+    import schnetpack.transform as trn
+
+    cutoff = 5.0
+    n_atom_basis = 32
+
+    pairwise_distance = (
+        spk.atomistic.PairwiseDistances()
+    )  # calculates pairwise distances between atoms
+    radial_basis = spk.nn.GaussianRBF(n_rbf=16, cutoff=cutoff)
+    painn = PaiNN(
+        n_atom_basis=n_atom_basis,
+        n_interactions=3,
+        radial_basis=radial_basis,
+        cutoff_fn=spk.nn.CosineCutoff(cutoff),
+    )
+    pred_U0 = spk.atomistic.Atomwise(n_in=n_atom_basis, output_key=QM9.U0)
+
+    nnpot = spk.model.NeuralNetworkPotential(
+        representation=painn,
+        input_modules=[pairwise_distance],
+        output_modules=[pred_U0],
+        postprocessors=[
+            trn.CastTo64(),
+            trn.AddOffsets(QM9.U0, add_mean=True, add_atomrefs=True),
+        ],
+    )
+
+    return nnpot
